pyproject.toml
[project]
name = "pdftoolkit"
version = "0.1.0"
description = "A collection of Python tools for extracting and analyzing PDF content, with a focus on charts and visualizations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "byaldi>=0.0.7",
    "docling>=2.66.0",
    "marker-pdf>=1.10.0",
    "markitdown[all]>=0.1.4",
    "megaparse>=0.0.55",
    "ollama>=0.6.0",
    "tenacity>=8.2.0",
    "together>=1.5.34",
    "unstructured[all-docs]>=0.16.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
pdftoolkit = "pdftoolkit.cli:main"
